+++ MoodPallette/sub/admin.py
@@ -1,0 +1,7 @@
+from django.contrib import admin
+
+from .models import subSet, subSelect
+
+# Register your models here.
+admin.site.register(subSet)
+admin.site.register(subSelect)
