--- MoodPallette/subSet/serializers.py
@@ -1,14 +1,0 @@
-from rest_framework import serializers
-from .models import subSet
-
-class subSetSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = subSet
-        fields = ('color', 'diffuser_name', 'handwash_name', 'handcream_name', 'sofrner_name', 'perfume_name', 'shampoo_name', 'rinse_name', 'bodywash_name', 'candle_name', 'room_name', 'bodymist_name', 'bodylotion_name')
-
-
-
-class NameSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = subSet
-        fields = ('diffuser_name', 'handwash_name', 'handcream_name', 'sofrner_name', 'perfume_name', 'shampoo_name', 'rinse_name', 'bodywash_name', 'candle_name', 'room_name', 'bodymist_name', 'bodylotion_name')
