--- MoodPallette/subSet/urls.py
@@ -1,9 +1,0 @@
-from django.contrib import admin
-from django.urls import path
-from subSet import views
-from django.views.decorators.csrf import csrf_exempt
-
-urlpatterns = [
-    path('<int:userId>/', csrf_exempt(views.subSet_API.as_view()))
-]
-
